Fixed bug with new application version checking mechanics in usage package
